merge

+++ src/common/FileSystem.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package common;
 
 import java.io.BufferedReader;
@@ -10,96 +5,134 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.FileReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.io.Serializable;
 import java.nio.channels.FileChannel;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
-/**
- *
- * @author red_f
- */
 public class FileSystem implements Serializable{
-    private String remoteHomeDir;
     private String localHomeDir;
     private String workingDirectory;
-    private String clientName;
-    private String currentServer;
+    private String ownerName;
+    //private String currentServer;
+    private String output;
     
     
     public FileSystem(String name){
-        this.clientName = name;
+        this.ownerName = name;
         //this.localHomeDir = "C:\\temp\\"+clientName;
-        this.localHomeDir = "C:/temp/";
-        this.workingDirectory = this.localHomeDir;
-        
-        
+        this.localHomeDir = "C:/temp";
+        this.workingDirectory = this.localHomeDir;   
     }
-    //monta vistas
-    public void buildTrees(){
-        
-        
-    }
-    
-    //gets
-    //sets
-    //fs comands
     
     public void Register(){}
     public void Login(){}
     public void Logout(){}
     
     //operators
-    public void copyFile(String fileName,String destinyPath){
+    public String copyFile(String fileName,String destinyPath){
+        output = "";
         FileChannel source;
         FileChannel destination;
-        if(this.workingDirectory.contains("remote")){
-            //request to server, get response;
+        
+        File s = new File(this.workingDirectory +"/" + fileName);
+        File d = new File(destinyPath +"/" +fileName);
+        try {
+           source = new FileInputStream(s).getChannel();
+           destination = new FileOutputStream(d).getChannel();
+           destination.transferFrom(source, 0, source.size());
+           source.close();
+           destination.close();
+           output+= "File copied to " +destinyPath;
+        } catch (FileNotFoundException ex) {
+            output += "File not found!";
+            return output;
+        } catch (IOException ex) {
+            output += "Data Error!";
+            return output;
         }
-        else{
-            File s = new File(this.workingDirectory + fileName);
-            File d = new File(destinyPath + fileName);
-            try {
-               source = new FileInputStream(s).getChannel();
-               destination = new FileOutputStream(d).getChannel();
-               destination.transferFrom(source, 0, source.size());
-               source.close();
-               destination.close();
-            } catch (FileNotFoundException ex) {
-                Logger.getLogger(FileSystem.class.getName()).log(Level.SEVERE, null, ex);
-            } catch (IOException ex) {
-                Logger.getLogger(FileSystem.class.getName()).log(Level.SEVERE, null, ex);
-            }
-        }
+        return output;
         
     }
-    public void moveFile(String fileName,String destinyPath){}
-    public void changeWorkingDirectory(String path){
-        this.workingDirectory = path;
-    }
-    public void getWorkingDirContent(){
-        if(this.workingDirectory.contains("remote")){
-            //request to server, get response;
-        }
-        else{
-            File folder = new File(this.workingDirectory);
-            File[] listOfFiles = folder.listFiles();
+    
+    public String moveFile(String fileName,String destinyPath){
+        output = "";
+        FileChannel source;
+        FileChannel destination;
+        
+        File s = new File(this.workingDirectory +"/" + fileName);
+        File d = new File(destinyPath +"/" +fileName);
+        try {
+            source = new FileInputStream(s).getChannel();
+            destination = new FileOutputStream(d).getChannel();
+            destination.transferFrom(source, 0, source.size());
+            source.close();
+            destination.close();
+            s.delete();
+            output = fileName +" moved to " + destinyPath;
 
-            for (int i = 0; i < listOfFiles.length; i++) {
-              if (listOfFiles[i].isFile()) {
-                System.out.println("[F]:" + listOfFiles[i].getName());
-              } else if (listOfFiles[i].isDirectory()) {
-                System.out.println("[D]:" + listOfFiles[i].getName());
-              }
+        } catch (FileNotFoundException ex) {
+            
+            output += "File not found!";
+            return output;
+        } catch (IOException ex) {
+            output += "Data Error!";
+            return output;
+        }
+        return output;
+    }
+    
+    public String changeWorkingDirectory(String path){
+        output ="";
+        
+        if(path.equalsIgnoreCase("CD..")){
+            //this.workingDirectory.
+            String[] npath = this.workingDirectory.split("/");
+            if(npath.length <= 1){
+                output+= "Working Directory is now " + this.workingDirectory + "/";
+                return output;
+            }
+            else{
+                
+                
+                this.workingDirectory = "";
+                for(int i = 0;i < npath.length-1;i++){
+                    
+                    this.workingDirectory+= npath[i] +"/";
+                }
+                output+= "Working Directory is now " + this.workingDirectory;
+                return output;
             }
         }
+        else{
+            //System.out.println("AQUI:" + path);
+            this.workingDirectory += "/" + path;
+            output += "Working Directory is now " + path + "/";
+            return output;
+        }
+    }
+    public String getDirContent(String fullPath) {
+        output="";
+       
+        File folder = new File(fullPath);
+        File[] listOfFiles = folder.listFiles();
         
+        if(listOfFiles != null){
+            output+=fullPath + "\n";
+            for (int i = 0; i < listOfFiles.length; i++) {
+                if (listOfFiles[i].isFile()) {
+                  output+="[F]:" + listOfFiles[i].getName() + "\n";
+                } else if (listOfFiles[i].isDirectory()) {
+                  output+="[D]:" + listOfFiles[i].getName() + "\n";
+                }
+            }
+            //output += "[]";
+        }else{
+            
+            output+=fullPath + "\n"+"empty:";
+        }
+        
+        return output;
         
         
     }
@@ -107,47 +140,63 @@ public class FileSystem implements Serializable{
         return this.workingDirectory;
     }
     //lel
-    public void getFileContent(){}
-    
-    public void removeFile(String fileName){
-        if(this.workingDirectory.contains("remote")){
-            //request to server, get response;
-        }
-        File f = new File(fileName);
-        f.delete();
-        System.out.println("Done!");
-    }
-    public void makeDir(String path){
-        if(this.workingDirectory.contains("remote")){
-            //request to server, get response;
-        }
-        File newDir = new File(path);
-        if(!newDir.exists()){
-            newDir.mkdir();
-            System.out.println("Done!");
-        }
-    }
-    
-    
-    //console debug only
-    public static void main(String[] args) throws IOException {
-        FileSystem fs = new FileSystem("red");
-        String msg;
-        BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
-        while(true){
-            System.out.print("@" + fs.workingDirectory +"> ");
-            msg = in.readLine();
-            switch(msg){
-                case "ls":
-                    fs.getWorkingDirContent();
-                    break;
+    public String getFileContent(String fileName){
+        output = "";
+        
+        try {
+            BufferedReader br = new BufferedReader(new FileReader(this.workingDirectory+"/"+fileName));
+            StringBuilder sb = new StringBuilder();
+            String line = br.readLine();
+
+            while (line != null) {
+                sb.append(line);
+                sb.append(System.lineSeparator());
+                line = br.readLine();
             }
+            output = sb.toString();
+            br.close();
+        } catch (FileNotFoundException ex) {
+            output+="File not found";
+        } catch (IOException ex) {
+            output+= "Error reading file!";
         }
+        /*File f = new File(this.workingDirectory + "/" + fileName);
+        try{
+            output+= Files.readAllLines(f.toPath());
+        } catch (IOException ex) {
+            output+= "Error reading file!";
+        }*/
+        return output;
+    }
+    
+    public String removeFile(String fileName){
+        output = "";
+        
+        File f = new File(this.workingDirectory + "/" +fileName);
+        if(f.exists()){
+            f.delete();
+            output += "Deleted " + this.workingDirectory + "/" +fileName + " sucessfuly!";
+        }else{
+            
+            output += "File not found!";
+        }
+        return output;
         
     }
-    
-    
-    
-    
-    
+    public String makeDir(String path){
+        
+        output ="";
+        File newDir = new File(this.workingDirectory + "/" + path);
+
+        if(!newDir.exists()){
+            newDir.mkdir();
+            output += this.workingDirectory + "/" + path + " created!";
+            
+        }
+        else{
+            output +="Directory already exists!";
+           
+        }
+        return output;
+    }   
 }
